refactor: split chain info from chains
