+++ src/index.tsx
@@ -71,4 +71,4 @@ ReactDOM.render(
 if (process.env.REACT_APP_SERVICE_WORKER !== 'false') {
   serviceWorkerRegistration.register()
 }
-export { INFURA_NETWORK_URLS } from './constants/chains'
+export { INFURA_NETWORK_URLS } from 'constants/chainInfo'
