+++ src/pages/Pool/CTACards.tsx
@@ -1,6 +1,7 @@
 import { Trans } from '@lingui/macro'
 import { AutoColumn } from 'components/Column'
-import { CHAIN_INFO, SupportedChainId } from 'constants/chains'
+import { CHAIN_INFO } from 'constants/chainInfo'
+import { SupportedChainId } from 'constants/chains'
 import { useActiveWeb3React } from 'hooks/web3'
 import styled from 'styled-components/macro'
 import { ThemedText } from 'theme'
