+++ src/utils/chains.tsx
@@ -1,4 +1,5 @@
-import { CHAIN_INFO, NetworkType, SupportedL1ChainId, SupportedL2ChainId } from '../constants/chains'
+import { CHAIN_INFO, NetworkType } from 'constants/chainInfo'
+import { SupportedL1ChainId, SupportedL2ChainId } from 'constants/chains'
 
 export function isL1ChainId(chainId: number | undefined): chainId is SupportedL1ChainId {
   return typeof chainId === 'number' && CHAIN_INFO[chainId].networkType === NetworkType.L1
